Fix warning sound on quit
