--- src/Pages/FrameworksPage.js
@@ -1,4 +1,0 @@
-const FrameworksPage = () => {
-  return <div>FrameworksPage</div>;
-};
-export default FrameworksPage;
