+++ src/components/App.js
@@ -2,7 +2,6 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
 import HomePage from '../pages/HomePage';
 import QuestionsHomePage from '../pages/QuestionsHomePage';
-import FrameworksPage from '../pages/FrameworksPage';
 import FinishedQuiz from 'pages/FinishedQuiz';
 
 import Questions from './Questions';
@@ -16,7 +15,7 @@ function App() {
           <Route path="/" element={<HomePage />} />
           <Route path="languages" element={<QuestionsHomePage />} />
           <Route path="questions/:language" element={<Questions />} />
-          <Route path="javascript/frameworks" element={<FrameworksPage />} />
+          {/* <Route path="javascript/frameworks" element={<FrameworksPage />} /> */}
           <Route path="score/:language" element={<FinishedQuiz />} />
           <Route path="*" element={<Error />} />
         </Routes>
