+++ src/components/IntroPictures.js
@@ -10,19 +10,6 @@ const IntroPictures = () => {
       <div className="intropic-container">
         {data.map((lang) => {
           const { img, name } = lang;
-
-          // Move to frameworks page
-          if (name === 'Javascript') {
-            return (
-              <Link to="/javascript/frameworks" key={name}>
-                <div className="picture-box">
-                  <img src={img} alt={name} className="intropage-img" />
-                </div>
-              </Link>
-            );
-          }
-
-          // category -medium, easy, hard, all
           return (
             <Link
               to={`/questions/${name}`}
