--- src/components/Main.js
@@ -1,4 +1,0 @@
-const Main = ({ children }) => {
-  return <main>{children}</main>;
-};
-export default Main;
