+++ src/data/questions.js
@@ -1,420 +1,3 @@
-export const JavascriptQuestions = {
-  Javascript: [
-    {
-      question: "What does the acronym 'DOM' stand for in JavaScript?",
-      options: [
-        'Document Object Model',
-        'Data Object Model',
-        'Document Order Model',
-        'Dynamic Object Model',
-      ],
-      correctAns: 0,
-      point: 10,
-    },
-    {
-      question:
-        'Which keywords enable block-scoped variable declaration in modern JavaScript?',
-      options: [
-        'var and const',
-        'let and var',
-        'let and const',
-        'const and var',
-      ],
-      correctAns: 2,
-      point: 10,
-    },
-    {
-      question: "What is the result of '5' + 3 in JavaScript?",
-      options: ['53', '8', '35', "'53'"],
-      correctAns: 3,
-      point: 10,
-    },
-    {
-      question:
-        'Which function is used to parse a string to an integer in JavaScript?',
-      options: ['parseInt()', 'toInteger()', 'parseString()', "'toInt()'"],
-      correctAns: 0,
-      point: 10,
-    },
-    {
-      question: "What does JavaScript's map() function do?",
-      options: [
-        'Create a new array by applying a function to each element.',
-        'To remove elements from an array.',
-        'To sort the elements in an array.',
-        'To filter out elements from an array.',
-      ],
-      correctAns: 0,
-      point: 20,
-    },
-    {
-      question:
-        "What is the purpose of the JavaScript 'addEventListener' method?",
-      options: [
-        'To remove an event listener from an element.',
-        'To add a new HTML element to the DOM.',
-        'To attach an event handler function to an HTML element.',
-        'To create a new event in JavaScript.',
-      ],
-      correctAns: 2,
-      point: 20,
-    },
-    {
-      question: 'What is the output of console.log(typeof NaN) in JavaScript?',
-      options: ["'number'", "'string'", "'NaN'", "'undefined'"],
-      correctAns: 0,
-      point: 20,
-    },
-    {
-      question: "What does the term 'callback hell' refer to in JavaScript?",
-      options: [
-        'Refers to excessive callback nesting, making code hard to read.',
-        'A type of function that always returns a callback function.',
-        'A popular JavaScript framework for handling callbacks.',
-        'A coding pattern for handling asynchronous code.',
-      ],
-      correctAns: 0,
-      point: 30,
-    },
-    {
-      question: 'What is the purpose of the JavaScript Promise object?',
-      options: [
-        'To create synchronous code in JavaScript.',
-        'Representing a future-resolved value.',
-        'To handle errors in asynchronous code.',
-        'To declare a variable with a fixed value.',
-      ],
-      correctAns: 1,
-      point: 30,
-    },
-    {
-      question: 'Do declarative functions hoist?',
-      options: [
-        "No, they don't.",
-        'Yes, they do',
-        "Only with 'var'.",
-        'They do, but after declaration.',
-      ],
-      correctAns: 1,
-      point: 30,
-    },
-  ],
-
-  React: [
-    {
-      question: 'Which company invented React?',
-      options: ['Google', 'Apple', 'Netflix', 'Facebook'],
-      correctOption: 3,
-      points: 10,
-    },
-    {
-      question:
-        "What's the name of the syntax we use to describe the UI in React components?",
-      options: ['FBJ', 'Babel', 'JSX', 'ES2015'],
-      correctOption: 2,
-      points: 10,
-    },
-    {
-      question: 'What is the virtual DOM in React?',
-      options: [
-        "A component that represents the browser's DOM.",
-        "A concept used to optimize React's rendering performance.",
-        'A method for querying the DOM using React.',
-        'A built-in HTML element in React.',
-      ],
-      correctOption: 1,
-      points: 10,
-    },
-    {
-      question: "What's the fundamental building block of React apps?",
-      options: ['Components', 'Blocks', 'Elements', 'Effects'],
-      correctOption: 0,
-      points: 10,
-    },
-    {
-      question: 'What is the purpose of state in a React component?',
-      options: [
-        "To store information that should not change during the component's lifetime.",
-        'To store data that can be modified and trigger re-renders when updated.',
-        "To define the component's layout and structure.",
-        'To manage routing in a React application.',
-      ],
-      correctOption: 1,
-      points: 20,
-    },
-    {
-      question:
-        'How can you pass data from a parent component to a child component in React?',
-      options: [
-        "Use the 'this.props' object.",
-        "Use the 'this.state' object.",
-        "Use the 'this.children' property.",
-        "Use the 'this.childProps' object.",
-      ],
-      correctOption: 1,
-      points: 20,
-    },
-    {
-      question:
-        'What is the purpose of the key prop when rendering lists of elements in React?',
-      options: [
-        'It is used to set the font size of list items.',
-        'It is used to uniquely identify each element in the list',
-        'It is used to specify the color of list items.',
-        'It is used to control the order of elements in the list.',
-      ],
-      correctOption: 1,
-      points: 20,
-    },
-
-    {
-      question: 'When do you use derived state?',
-      options: [
-        'Whenever the state should not trigger a re-render',
-        'Whenever the state can be synchronized with an effect',
-        'Whenever the state should be accessible to all components',
-        'Whenever the state can be computed from another state variable',
-      ],
-      correctOption: 3,
-      points: 30,
-    },
-    {
-      question: 'In what situation do we use a callback to update state?',
-      options: [
-        'When updating the state will be slow',
-        'When the updated state is very data-intensive',
-        'When the state update should happen faster',
-        'When the new state depends on the previous state',
-      ],
-      correctOption: 3,
-      points: 30,
-    },
-    {
-      question:
-        "When will an effect run if it doesn't have a dependency array?",
-      options: [
-        'Only when the component mounts',
-        'Only when the component unmounts',
-        'The first time the component re-renders',
-        'Each time the component is re-rendered',
-      ],
-      correctOption: 3,
-      points: 30,
-    },
-  ],
-
-  Angular: [
-    {
-      question: 'What is the primary building block of an Angular application?',
-      options: ['Functions', 'Modules', 'HTML', 'CSS'],
-      correctOption: 1,
-      points: 10,
-    },
-    {
-      question:
-        'How do you create a new Angular component using the Angular CLI?',
-      options: [
-        'ng new component',
-        'ng create component',
-        'ng generate component',
-        'ng add component',
-      ],
-      correctOption: 2,
-      points: 10,
-    },
-    {
-      question: 'In Angular, what is the purpose of a service?',
-      options: [
-        'To define the UI structure of a component',
-        "To manage the application's state",
-        'To style the user interface',
-        'To handle HTTP requests and share data between components',
-      ],
-      correctOption: 3,
-      points: 10,
-    },
-    {
-      question:
-        'Which directive in Angular is used to conditionally display or hide an element in the template?',
-      options: ['*ngIf', '*ngFor', '*ngSwitch', '*ngInclude'],
-      correctOption: 0,
-      points: 10,
-    },
-    {
-      question: 'How do you handle user input in Angular forms?',
-      options: [
-        'Use the @Input decorator.',
-        'Use the @Output decorator.',
-        'Use the (ngModel) directive.',
-        'Use the (click) event binding.',
-      ],
-      correctOption: 2,
-      points: 20,
-    },
-    {
-      question: 'What is lazy loading in the context of Angular?',
-      options: [
-        'Loading Angular modules asynchronously.',
-        'Loading Angular components one by one.',
-        'Delaying the rendering of Angular templates.',
-        'Loading Angular apps only when the user explicitly requests them.',
-      ],
-      correctOption: 0,
-      points: 20,
-    },
-    {
-      question: 'What is the purpose of the Angular RouterModule?',
-      options: [
-        'To create animations in Angular components',
-        'To manage HTTP requests',
-        'To handle routing and navigation within the Angular application',
-        'To create custom directives',
-      ],
-      correctOption: 2,
-      points: 20,
-    },
-    {
-      question: 'How can you share data between sibling components in Angular?',
-      options: [
-        'Use input properties.',
-        'Use output properties and event emitters.',
-        'Use a shared service or a state management library like NgRx or Redux.',
-        'Use template reference variables.',
-      ],
-      correctOption: 2,
-      points: 30,
-    },
-    {
-      question:
-        "In Angular, what is the purpose of the '@ViewChild' decorator?",
-      options: [
-        'To inject a service into a component',
-        "To access a child component or element in the component's template",
-        'To define a variable within a component',
-        'To create a new instance of a component',
-      ],
-      correctOption: 1,
-      points: 30,
-    },
-    {
-      question: "What is Angular's HTTP client module used for?",
-      options: [
-        'To manage Angular components',
-        'To create and manage HTTP requests',
-        'To handle user authentication',
-        'To render Angular templates',
-      ],
-      correctOption: 1,
-      points: 30,
-    },
-  ],
-
-  Vue: [
-    {
-      question: 'What is the core library of Vue.js called?',
-      options: ['Vue', ' VueJS', ' VueCore', ' VueLib'],
-      correctOption: 0,
-      points: 10,
-    },
-    {
-      question: 'How do you bind data to the DOM in Vue.js?',
-      options: [
-        'Using Angular directives',
-        'Using React components',
-        'Using Vue directives',
-        'Using jQuery',
-      ],
-      correctOption: 2,
-      points: 10,
-    },
-    {
-      question: 'What is a Vue instance?',
-      options: [
-        ' An instance of a class in Vue.js',
-        'A JavaScript object created from a Vue constructor',
-        'A server running Vue applications',
-        'A Vue component',
-      ],
-      correctOption: 1,
-      points: 10,
-    },
-    {
-      question: 'What is the template syntax in Vue.js used for?',
-      options: [
-        'Defining CSS styles',
-        'Defining routes in a Vue application',
-        'Defining the structure of the user interface',
-        'Defining API endpoints',
-      ],
-      correctOption: 2,
-      points: 10,
-    },
-    {
-      question:
-        "What is Vue's official state management library for managing application-level data?",
-      options: ['Redux', 'Flux', 'Vuex', 'VueX'],
-      correctOption: 2,
-      points: 20,
-    },
-    {
-      question: 'What is Vue Router used for in Vue.js applications?',
-      options: [
-        'Handling HTTP requests',
-        "Managing the application's state",
-        'Managing client-side routing and navigation',
-        'Defining server routes',
-      ],
-      correctOption: 2,
-      points: 20,
-    },
-    {
-      question: 'What is a Vue component?',
-      options: [
-        'A standalone Vue instance',
-        'A custom HTML element in Vue.js',
-        'A built-in HTML element in Vue.js',
-        'A Vue directive',
-      ],
-      correctOption: 1,
-      points: 20,
-    },
-    {
-      question: 'What does the v-model directive in Vue.js allow you to do?',
-      options: [
-        'Create custom directives',
-        "Bind an input element's value to a component's data",
-        'Define routes and navigation in Vue applications',
-        'Implement two-way data binding',
-      ],
-      correctOption: 3,
-      points: 30,
-    },
-    {
-      question: 'What is the Vue.js CLI used for?',
-      options: [
-        'Managing state in Vue applications',
-        'Creating and managing Vue projects, including scaffolding and development tools',
-        'Routing in Vue applications',
-        'Defining custom Vue directives',
-      ],
-      correctOption: 1,
-      points: 30,
-    },
-    {
-      question:
-        'What is the purpose of Vue.js transition and animation classes?',
-      options: [
-        'They are used for defining API endpoints in Vue applications.',
-        'They are used to create custom directives.',
-        'They are used to animate Vue components when they enter or leave the DOM.',
-        'They are used for managing server-side rendering (SSR).',
-      ],
-      correctOption: 2,
-      points: 30,
-    },
-  ],
-};
-
 export const languagesQuestions = {
   'C++': [
     {
@@ -970,7 +553,422 @@ export const languagesQuestions = {
         'A stack of deferred functions executed in a specific order, often used for cleanup.',
       ],
       correctOption: 2,
+      points: 30,
+    },
+  ],
+
+  Javascript: [
+    {
+      question: "What does the acronym 'DOM' stand for in JavaScript?",
+      options: [
+        'Document Object Model',
+        'Data Object Model',
+        'Document Order Model',
+        'Dynamic Object Model',
+      ],
+      correctAns: 0,
+      points: 10,
+    },
+    {
+      question:
+        'Which keywords enable block-scoped variable declaration in modern JavaScript?',
+      options: [
+        'var and const',
+        'let and var',
+        'let and const',
+        'const and var',
+      ],
+      correctAns: 2,
+      points: 10,
+    },
+    {
+      question: "What is the result of '5' + 3 in JavaScript?",
+      options: ['53', '8', '35', "'53'"],
+      correctAns: 3,
+      points: 10,
+    },
+    {
+      question:
+        'Which function is used to parse a string to an integer in JavaScript?',
+      options: ['parseInt()', 'toInteger()', 'parseString()', "'toInt()'"],
+      correctAns: 0,
+      points: 10,
+    },
+    {
+      question: "What does JavaScript's map() function do?",
+      options: [
+        'Create a new array by applying a function to each element.',
+        'To remove elements from an array.',
+        'To sort the elements in an array.',
+        'To filter out elements from an array.',
+      ],
+      correctAns: 0,
       points: 20,
+    },
+    {
+      question:
+        "What is the purpose of the JavaScript 'addEventListener' method?",
+      options: [
+        'To remove an event listener from an element.',
+        'To add a new HTML element to the DOM.',
+        'To attach an event handler function to an HTML element.',
+        'To create a new event in JavaScript.',
+      ],
+      correctAns: 2,
+      points: 20,
+    },
+    {
+      question: 'What is the output of console.log(typeof NaN) in JavaScript?',
+      options: ["'number'", "'string'", "'NaN'", "'undefined'"],
+      correctAns: 0,
+      points: 20,
+    },
+    {
+      question: "What does the term 'callback hell' refer to in JavaScript?",
+      options: [
+        'Refers to excessive callback nesting, making code hard to read.',
+        'A type of function that always returns a callback function.',
+        'A popular JavaScript framework for handling callbacks.',
+        'A coding pattern for handling asynchronous code.',
+      ],
+      correctAns: 0,
+      points: 30,
+    },
+    {
+      question: 'What is the purpose of the JavaScript Promise object?',
+      options: [
+        'To create synchronous code in JavaScript.',
+        'Representing a future-resolved value.',
+        'To handle errors in asynchronous code.',
+        'To declare a variable with a fixed value.',
+      ],
+      correctAns: 1,
+      points: 30,
+    },
+    {
+      question: 'Do declarative functions hoist?',
+      options: [
+        "No, they don't.",
+        'Yes, they do',
+        "Only with 'var'.",
+        'They do, but after declaration.',
+      ],
+      correctAns: 1,
+      points: 30,
+    },
+  ],
+
+  React: [
+    {
+      question: 'Which company invented React?',
+      options: ['Google', 'Apple', 'Netflix', 'Facebook'],
+      correctOption: 3,
+      points: 10,
+    },
+    {
+      question:
+        "What's the name of the syntax we use to describe the UI in React components?",
+      options: ['FBJ', 'Babel', 'JSX', 'ES2015'],
+      correctOption: 2,
+      points: 10,
+    },
+    {
+      question: 'What is the virtual DOM in React?',
+      options: [
+        "A component that represents the browser's DOM.",
+        "A concept used to optimize React's rendering performance.",
+        'A method for querying the DOM using React.',
+        'A built-in HTML element in React.',
+      ],
+      correctOption: 1,
+      points: 10,
+    },
+    {
+      question: "What's the fundamental building block of React apps?",
+      options: ['Components', 'Blocks', 'Elements', 'Effects'],
+      correctOption: 0,
+      points: 10,
+    },
+    {
+      question: 'What is the purpose of state in a React component?',
+      options: [
+        "To store information that should not change during the component's lifetime.",
+        'To store data that can be modified and trigger re-renders when updated.',
+        "To define the component's layout and structure.",
+        'To manage routing in a React application.',
+      ],
+      correctOption: 1,
+      points: 20,
+    },
+    {
+      question:
+        'How can you pass data from a parent component to a child component in React?',
+      options: [
+        "Use the 'this.props' object.",
+        "Use the 'this.state' object.",
+        "Use the 'this.children' property.",
+        "Use the 'this.childProps' object.",
+      ],
+      correctOption: 1,
+      points: 20,
+    },
+    {
+      question:
+        'What is the purpose of the key prop when rendering lists of elements in React?',
+      options: [
+        'It is used to set the font size of list items.',
+        'It is used to uniquely identify each element in the list',
+        'It is used to specify the color of list items.',
+        'It is used to control the order of elements in the list.',
+      ],
+      correctOption: 1,
+      points: 20,
+    },
+
+    {
+      question: 'When do you use derived state?',
+      options: [
+        'Whenever the state should not trigger a re-render',
+        'Whenever the state can be synchronized with an effect',
+        'Whenever the state should be accessible to all components',
+        'Whenever the state can be computed from another state variable',
+      ],
+      correctOption: 3,
+      points: 30,
+    },
+    {
+      question: 'In what situation do we use a callback to update state?',
+      options: [
+        'When updating the state will be slow',
+        'When the updated state is very data-intensive',
+        'When the state update should happen faster',
+        'When the new state depends on the previous state',
+      ],
+      correctOption: 3,
+      points: 30,
+    },
+    {
+      question:
+        "When will an effect run if it doesn't have a dependency array?",
+      options: [
+        'Only when the component mounts',
+        'Only when the component unmounts',
+        'The first time the component re-renders',
+        'Each time the component is re-rendered',
+      ],
+      correctOption: 3,
+      points: 30,
+    },
+  ],
+
+  Angular: [
+    {
+      question: 'What is the primary building block of an Angular application?',
+      options: ['Functions', 'Modules', 'HTML', 'CSS'],
+      correctOption: 1,
+      points: 10,
+    },
+    {
+      question:
+        'How do you create a new Angular component using the Angular CLI?',
+      options: [
+        'ng new component',
+        'ng create component',
+        'ng generate component',
+        'ng add component',
+      ],
+      correctOption: 2,
+      points: 10,
+    },
+    {
+      question: 'In Angular, what is the purpose of a service?',
+      options: [
+        'To define the UI structure of a component',
+        "To manage the application's state",
+        'To style the user interface',
+        'To handle HTTP requests and share data between components',
+      ],
+      correctOption: 3,
+      points: 10,
+    },
+    {
+      question:
+        'Which directive in Angular is used to conditionally display or hide an element in the template?',
+      options: ['*ngIf', '*ngFor', '*ngSwitch', '*ngInclude'],
+      correctOption: 0,
+      points: 10,
+    },
+    {
+      question: 'How do you handle user input in Angular forms?',
+      options: [
+        'Use the @Input decorator.',
+        'Use the @Output decorator.',
+        'Use the (ngModel) directive.',
+        'Use the (click) event binding.',
+      ],
+      correctOption: 2,
+      points: 20,
+    },
+    {
+      question: 'What is lazy loading in the context of Angular?',
+      options: [
+        'Loading Angular modules asynchronously.',
+        'Loading Angular components one by one.',
+        'Delaying the rendering of Angular templates.',
+        'Loading Angular apps only when the user explicitly requests them.',
+      ],
+      correctOption: 0,
+      points: 20,
+    },
+    {
+      question: 'What is the purpose of the Angular RouterModule?',
+      options: [
+        'To create animations in Angular components',
+        'To manage HTTP requests',
+        'To handle routing and navigation within the Angular application',
+        'To create custom directives',
+      ],
+      correctOption: 2,
+      points: 20,
+    },
+    {
+      question: 'How can you share data between sibling components in Angular?',
+      options: [
+        'Use input properties.',
+        'Use output properties and event emitters.',
+        'Use a shared service or a state management library like NgRx or Redux.',
+        'Use template reference variables.',
+      ],
+      correctOption: 2,
+      points: 30,
+    },
+    {
+      question:
+        "In Angular, what is the purpose of the '@ViewChild' decorator?",
+      options: [
+        'To inject a service into a component',
+        "To access a child component or element in the component's template",
+        'To define a variable within a component',
+        'To create a new instance of a component',
+      ],
+      correctOption: 1,
+      points: 30,
+    },
+    {
+      question: "What is Angular's HTTP client module used for?",
+      options: [
+        'To manage Angular components',
+        'To create and manage HTTP requests',
+        'To handle user authentication',
+        'To render Angular templates',
+      ],
+      correctOption: 1,
+      points: 30,
+    },
+  ],
+
+  Vue: [
+    {
+      question: 'What is the core library of Vue.js called?',
+      options: ['Vue', ' VueJS', ' VueCore', ' VueLib'],
+      correctOption: 0,
+      points: 10,
+    },
+    {
+      question: 'How do you bind data to the DOM in Vue.js?',
+      options: [
+        'Using Angular directives',
+        'Using React components',
+        'Using Vue directives',
+        'Using jQuery',
+      ],
+      correctOption: 2,
+      points: 10,
+    },
+    {
+      question: 'What is a Vue instance?',
+      options: [
+        ' An instance of a class in Vue.js',
+        'A JavaScript object created from a Vue constructor',
+        'A server running Vue applications',
+        'A Vue component',
+      ],
+      correctOption: 1,
+      points: 10,
+    },
+    {
+      question: 'What is the template syntax in Vue.js used for?',
+      options: [
+        'Defining CSS styles',
+        'Defining routes in a Vue application',
+        'Defining the structure of the user interface',
+        'Defining API endpoints',
+      ],
+      correctOption: 2,
+      points: 10,
+    },
+    {
+      question:
+        "What is Vue's official state management library for managing application-level data?",
+      options: ['Redux', 'Flux', 'Vuex', 'VueX'],
+      correctOption: 2,
+      points: 20,
+    },
+    {
+      question: 'What is Vue Router used for in Vue.js applications?',
+      options: [
+        'Handling HTTP requests',
+        "Managing the application's state",
+        'Managing client-side routing and navigation',
+        'Defining server routes',
+      ],
+      correctOption: 2,
+      points: 20,
+    },
+    {
+      question: 'What is a Vue component?',
+      options: [
+        'A standalone Vue instance',
+        'A custom HTML element in Vue.js',
+        'A built-in HTML element in Vue.js',
+        'A Vue directive',
+      ],
+      correctOption: 1,
+      points: 20,
+    },
+    {
+      question: 'What does the v-model directive in Vue.js allow you to do?',
+      options: [
+        'Create custom directives',
+        "Bind an input element's value to a component's data",
+        'Define routes and navigation in Vue applications',
+        'Implement two-way data binding',
+      ],
+      correctOption: 3,
+      points: 30,
+    },
+    {
+      question: 'What is the Vue.js CLI used for?',
+      options: [
+        'Managing state in Vue applications',
+        'Creating and managing Vue projects, including scaffolding and development tools',
+        'Routing in Vue applications',
+        'Defining custom Vue directives',
+      ],
+      correctOption: 1,
+      points: 30,
+    },
+    {
+      question:
+        'What is the purpose of Vue.js transition and animation classes?',
+      options: [
+        'They are used for defining API endpoints in Vue applications.',
+        'They are used to create custom directives.',
+        'They are used to animate Vue components when they enter or leave the DOM.',
+        'They are used for managing server-side rendering (SSR).',
+      ],
+      correctOption: 2,
+      points: 30,
     },
   ],
 };
